extract umi matrix for the genes in the set, possibly donwsampling

extract umi matrix for the genes in the set, possibly donwsampling

gset_get_feat_mat(
  gset_id,
  mat_id,
  downsamp = F,
  add_non_dsamp = F,
  downsample_n = NA,
  gene_names_src_targ = NULL
)

Arguments

gset_id

gene set in scdb

mat_id

mat in scdb

downsamp

if this is true the returned matrix is downsampled

downsample_n

number of umis for downsampling (if this is NA, the system will pick up the default heuristic)

gene_names_src_targ

string pair, defining how to convert gene nmaes from the original matrix to the naming scheme of the returned matrix (NA by default, in which case no conversion will be done)

add_non+dsamp

should cells with fewer umis than the downsampling threshold be added to hte matrix with their raw data