colorize metacells using a set of super MCs derived by hclust, colored according to a user defined table

colorize metacells using a set of super MCs derived by hclust, colored according to a user defined table

mc_colorize_sup_hierarchy(mc_id, supmc, supmc_key, gene_key = NULL)

Arguments

mc_id
metacell id in scdb

supmc
output from mcell_mc_hierarchy, defining groups of mcs

supmc_key
filename of color index, specifying an ordered list of supmc ids and colors, to be colored in the order of apperance (i.e. last line overrid previous lines). Expected field names: supid, color, name