Annotate query metacell with atlas by comparing query cells to atlas MCs

This will take each cell in the query MC and find its to correlated metacell in the reference, then generating figures showing detailed comparison of how each metacell in the query is distributed in the atlas, and how the pool of all cells in the query MC compare to the pool of their best match projections

mcatlas_annotate_mc_by_sc2mc_projection(
  atlas_id,
  qmat_id,
  qmc_id,
  qmat_naming_type,
  fig_cmp_dir = NULL,
  q_gset_id = NULL,
  new_qmc_id = NULL,
  atlas_mc2d_id = NULL,
  plot_all_mcs = F,
  md_field = NULL,
  max_entropy = 2,
  burn_cor = 0.6
)

Arguments

atlas_id
id of atlas object in scdb

qmat_id
id of metacell object ina scdb

qmc_id
id of metacell object ina scdb

qmat_naming_type
naming scheme of query matrix/mc

fig_cmp_dir
name of directory to put figures per MC

q_gset_id
query gene set id object (optional) - to restrict features used in comaprison to the intersection of atlas and query gsets

md_field
metadata field too use as additional factor for plotting subsets

recolor_mc_id
if this is specified, the atlas colors will be projected on the query MCCs and updated to the scdb object named recolor_mc

plot_all_mc
set this to T if you want a plot per metacell to show comparison of query pooled umi's and projected pooled umis.