Annotate query metacell with atlas by comparing query cells to altas cells with a Knn strategy Algorithm short description TBA

Annotate query metacell with atlas by comparing query cells to altas cells with a Knn strategy Algorithm short description TBA

mcatlas_annotate_mc_by_sc2sc_projection(
  atlas_id,
  qmc_id,
  qmat_id,
  qmat_naming_type,
  all_vs_all,
  new_qmc_id,
  T_cor_gap = 1,
  fig_cmp_dir = NULL,
  sc_annots = NULL
)

Arguments

atlas_id
id of atlas object in scdb

qmc_id
id of metacell object ina scdb

qmat_id
id of query umi matrix

qmat_naming_type
naming scheme of query matrix/mc

all_vs_all
should a graph combining query and atlas be constructed, or only K-nn of a bipartiate query atlas graph be examined.

new_qmc_id
id of recolored metacell object to save in DB (NULL will supress updating db)

T_cor_gap
how much gap between internal atlas mc-mc correlation and query-atlas correlation one is allowing to keep the annotation.

fig_cmp_dir
name of directory to put figures per MC

sc_annot
cached sc_annot object can be specified (as returned by mcatlas_annotate_sc_by_projection) - to save time in repeated annotation refinement