Compute predictive value of MC cover as correlation of MC averages of sc umis

Compute predictive value of MC cover as correlation of MC averages of sc umis

mcell_graph_pred_corr(
  mat_id,
  graph_id,
  do_log = F,
  focus_cells = NULL,
  filt_genes = NULL
)

Arguments

mat_id
id of matrix to analyze

graph_id
id of cgraph to use for computing neighborhood per cell

do_log
should log transofmration be performed

focus_cells
should correltion be computed to a specific set of cells