Metacell layout using force directed projection of a low degree mc graph

Metacell layout using force directed projection of a low degree mc graph

mcell_mc2d_force_knn(
  mc2d_id,
  mc_id,
  graph_id,
  mgraph_id = NULL,
  ignore_mismatch = F,
  symmetrize = F,
  ignore_edges = NULL,
  feats_gset = NULL,
  feats_exclude = NULL,
  graph_parametric = F,
  logist_loc = 1,
  logist_scale = 0.2,
  logist_eps = 0.00004,
  max_d_fold = 3,
  use_umap_for_2d_pos = F,
  umap_mgraph = F,
  uconf = NULL
)

Arguments

mc2d_id

2d object to add

mc_id

meta cell id to work with

graph_id

graph_id of the similarity graph on cells from the metacell. This should always be specified, even if an mgraph is supplied since we use the obejct to position cells releative to metacels

mgraph_id

mgraph_id of the metacell graph to use. If this is null the function will compute an mgraph itself using the cell graph

ignore_mismatch

determine if mismatch in mc names are to be ignored.

feats_gset

gene set name for use in parametric graph and/or umap 2d projection

feats_exclude

list of genes to exclude from the features gene set

graph_parametric

if this is True, mc edges will be determined using parametric distances and not the cells k-nn graph.

logist_loc

the location parametr of the logistic function used to determine parametric distances between metacelles

logist_scale

the lscale parametr of the logistic function used to determine parametric distances between metacelles

symetrize

should the mc confusion matrix be symmetrized before computing layout?