Utility plot function to compare two genes based on mc_fp

Utility plot function to compare two genes based on mc_fp

mcell_mc_plot_gg(
  mc_id,
  g1,
  g2,
  cex = 2,
  text_cex = 0.5,
  mc_colors = NULL,
  fig_fn = NULL,
  cell_md = NULL,
  md_mode = "filter",
  use_egc = F,
  main = NULL,
  cex.main = 1,
  xlim = NULL,
  ylim = NULL,
  add_grid = F,
  e_gc_eps = 0.00001,
  mc_filt = NULL,
  md = NULL,
  md_regexp = NULL,
  md_field = NULL
)

Arguments

mc_id

metacell id in scdb

g1

gene name

g2

gene name

cex

point size

text_cex

text size (default 0.5)

mc_colors

coloring vector per MC, default is NULL in which case we use the MC colors

fig_fn

file name for the figure (optional - default is current device)

cell_md

a factor/logical vector defining which cells are in the focus. If this is unspecified all cells/metacells will be included

md_mode

one of "filter", "highlight"

use_egc

plot log of absolute expression instead of the log of relative expression (normalized to the median over all MCs)

mc_filt

potential facator filtering metacell ids (default NULL)

md

metadata to use for highlighting/filtering

md_regexp

regular expression to filter/highlight metadata

md_field

metadata field to use