plot a marker heat map give a metacell object

plot a marker heat map give a metacell object

mcell_mc_plot_marks(
  mc_id,
  gset_id,
  mat_id = mc_id,
  fig_fn = NULL,
  lateral_gset_id = NULL,
  mc_ord = NULL,
  plot_cells = T,
  zero_median = T,
  add_genes = NULL,
  add_metadata = NULL,
  ext_metadata = NULL,
  md_level_colors = NULL,
  focus_mcs = NULL,
  gene_list = NULL,
  reorder_marks = F,
  fold_burn = 3
)

Arguments

mc_id

id of metacell object ina scdb

gset_id

markers to plot (e.g. using mcell_gset_from_mc_markers)

mat_id

matrix object to us (default is the mc_id - assuming they use the same ID), if some cells in mc@mc are missing from the matrix, the function will generate an error

fig_fn

file name for the figure (if null it will be call heat_marks in the fig directory)

lateral_gset_id

markers to plot on top (e.g. cell cycle)

plot_cells

by defulat this is TRUE and data is shown for single cells. If this is false, than metacells fold change values will be plotted

zero_median

if this is true then for each gene, all values below the median will be plotted as "0". It can save some noise but will make genes that are anit-enriched in some MC non visible.

add_genes

A list of marker genes to be added manually

add_metadata

name of metadata field to add below hte heat map

ext_metadata

values of metadata per cell to add (a anamed vector, with some or all of the cell names specified)

md_level_colors

what colors to use for the metadata levels, if present

focus_mcs

option list of MC IDs in case only some of the model should be plotted

gene_list

a list of gene names to replcace the markers from a gset_id