Calculate basic statistics on a matrix

Calculate basic statistics on a matrix

scm_gene_stat(
  mat_id,
  niche_quantile = 0.2,
  downsample_n = NULL,
  K_std_n = 1000
)

Arguments

niche_quantile

A value between 0 and 1.

mat

The input matrix

Value

Returns a dataset that contains statistic for all genes that passed the filtering stage. Columns starting with ds contain UMI statistics after downsampling, columns starting with n contain UMI statistics after normalizing UMIs so that the number of UMIs per cell sums to 1. The columns are:

tot

Total gene expression

var

Gene variance

is_on_count

Number of cells in which the gene is expressed

sz_cor

Correlation with cell size

sz_cor_norm

sz_cor after subtracting the trend

niche_stat

How many of the genen's umis are found in X% of the most highly expressing cells. (regularized)

niche_norm

niche_stat after subtracting the niche_norm trend: median niche_norm value of genes with similar total expression

n_mean

Mean after normalization

ds_top1

Largest count, after downsampling

ds_top2

2nd largest count, after downsampling

ds_top3

3rd largest count, after downsampling

ds_mean

Mean on downsampled data

ds_var

Variance on downsampled data

ds_log_varmean

log2 of ds_var/ds_mean

ds_vm_norm

ds_log_varmean after subtracting the trend

ds_is_on_count

Number of cells in which the gene is expressed, after down sampling

downsample_n

Number of UMIs used for downsampling