Single cell RNA-seq matrix

a single cell RNA matrix interface. Does not do much beyond adding specifically gene names, cell names and labels describing the type of data. Methods for basic stat, gene selection, clustering are implemented separately

Slots

mat

A sparse matrix containing the expression data

genes

list of all genes. By default the matrix row names are used

cells

list of all cells. By default the matrix col names are used

stat_type

Type of statistic. Currently only "umi" is supported

cell_metadata

dataframe with metadata on the cells. Rownames correspond to the cell names. By default a table consisting of only a "Batch" field is added with a trivial batch number. If the cell_metadata does not include a Batch field, it will be added with a trivial value. Additional standard fields spike_umis - the number of (filtered_) spike umis for the cell