Utility plot function to compare bulk expression of two batches/metadata factors

Utility plot function to compare bulk expression of two batches/metadata factors

mcell_compare_bulk(
  mat_id,
  meta_field = NULL,
  v1 = NULL,
  v2 = NULL,
  mc_id = NULL,
  mc1 = -1,
  mc2 = -1,
  mc_to_focus = NULL,
  genes = NULL,
  cex = 0.5,
  text_cex = 1,
  n_reg = 5,
  fig_nm = NULL,
  top_genes = 10,
  fig_h = 600,
  fig_w = 800
)

Arguments

mat_id
matrix id to use

meta_field
metadata field to select cells, if not specified than a metacell must be specified

v1
first set metadata value

v2
secod set metadata value

mc_id
metacell model (must be specified if meta_field is not used)

mc1
metacell id to define cells in first set

mc2
metacell id to define cells in second set

mc_to_focus
metacell IDs to restrict analysis of metadata comparison

genes
genes to highlight (optional)